requirements.txt
atomicwrites==1.4.0
attrs==20.3.0
aws-cdk.assets==1.75.0
aws-cdk.aws-applicationautoscaling==1.75.0
aws-cdk.aws-autoscaling-common==1.75.0
aws-cdk.aws-cloudwatch==1.75.0
aws-cdk.aws-codeguruprofiler==1.75.0
aws-cdk.aws-ec2==1.75.0
aws-cdk.aws-efs==1.75.0
aws-cdk.aws-events==1.75.0
aws-cdk.aws-iam==1.75.0
aws-cdk.aws-kms==1.75.0
aws-cdk.aws-lambda==1.75.0
aws-cdk.aws-logs==1.75.0
aws-cdk.aws-s3==1.75.0
aws-cdk.aws-s3-assets==1.75.0
aws-cdk.aws-sns==1.75.0
aws-cdk.aws-sns-subscriptions==1.75.0
aws-cdk.aws-sqs==1.75.0
aws-cdk.aws-ssm==1.75.0
aws-cdk.cloud-assembly-schema==1.75.0
aws-cdk.core==1.75.0
aws-cdk.cx-api==1.75.0
aws-cdk.region-info==1.75.0
cattrs==1.1.1
certifi==2020.11.8
chardet==3.0.4
colorama==0.4.4
constructs==3.2.40
cssselect==1.1.0
GetOldTweets3==0.0.11
idna==2.10
iniconfig==1.1.1
jsii==1.15.0
lxml==4.6.2
markovify==0.8.3
packaging==20.4
pluggy==0.13.1
publication==0.0.3
py==1.9.0
pyparsing==2.4.7
pyquery==1.4.3
pytest==6.1.2
python-dateutil==2.8.1
requests==2.25.0
six==1.15.0
toml==0.10.2
typing-extensions==3.7.4.3
Unidecode==1.1.1
urllib3==1.26.2
